moved new commands under the internal/cli/command package
