+++ internal/cli/internal/command/apps/destroy.go
