+++ internal/cli/internal/command/apps/list.go
