+++ internal/cli/internal/command/apps/restart.go
