+++ internal/cli/internal/command/apps/suspend.go
