+++ internal/cli/internal/command/version/init_state.go
